Moves fn validateFeatureQuery to map_functions file
